- adds the description copy upon migration
