+++ src/kiota/Handlers/Client/RemoveHandler.cs
@@ -29,7 +29,7 @@ internal class RemoveHandler : BaseKiotaCommandHandler
             try
             {
                 await CheckForNewVersionAsync(logger, cancellationToken).ConfigureAwait(false);
-                var workspaceManagementService = new WorkspaceManagementService(logger, true);
+                var workspaceManagementService = new WorkspaceManagementService(logger, httpClient, true);
                 await workspaceManagementService.RemoveClientAsync(className, cleanOutput, cancellationToken).ConfigureAwait(false);
                 return 0;
             }
